requirements.txt
# Veloquity MVP — Python dependencies

# API
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.2

# Database
psycopg2-binary==2.9.9
pgvector==0.2.5

# AWS SDK
boto3==1.34.0

# Utilities
python-dotenv==1.0.1
